confirm and busy spinner to comment cards

+++ client/src/modules/comments/comment-card/comment-card.js
@@ -5,17 +5,28 @@ import template from './comment-card.html';
 import './comment-card.css';
 
 class CommentCardController {
-  constructor(UserService, MomentService, EventService, GroupService) {
+  constructor(UserService, MomentService, EventService, GroupService, ConfirmService) {
     this.user = UserService.user;
     this.moment = MomentService.moment;
     this.EventService = EventService;
     this.groupOwner = GroupService.currentGroup.userId;
+    this.ConfirmService = ConfirmService;
+
+    this.busy = false;
+
     this.handleDeleteClick = this.handleDeleteClick.bind(this);
     this.menuShouldAppear = this.menuShouldAppear.bind(this);
   }
 
   handleDeleteClick() {
-    this.EventService.removeCommentForEvent(this.eventId, this.comment._id);
+    this.ConfirmService.openModal(
+      'Are you sure you want to delete this comment?',
+      'This action cannot be undone', 'Yes'
+    ).then(() => {
+      this.busy = true;
+      this.EventService.removeCommentForEvent(this.eventId, this.comment._id)
+        .finally(() => this.busy = false);
+    });
   }
 
   menuShouldAppear() {
@@ -24,7 +35,10 @@ class CommentCardController {
   }
 
 }
-CommentCardController.$inject = ['UserService', 'MomentService', 'EventService', 'GroupService'];
+CommentCardController.$inject = [
+  'UserService', 'MomentService', 'EventService',
+  'GroupService', 'ConfirmService'
+];
 
 const CommentCardComponent = {
   restrict: 'E',
